Make healthcheck drain time configurable
Increasing this time to 20 seconds by default allows load balancers
that UAA may be behind to unregister the instance. We make it configurable
because load balancer's configurations may vary.

Without this extra drain time the UAA API will suffer downtime of a few
seconds when being restarted or redeployed, which is unacceptable in a
high availability deployment.

The knock-on effect of increasing this script's execution time is we
have to increase the Monit stop operation timeout.